Add the application class in which the DbManager is created for use by project classes.
