+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     package="com.example.currentnote">
 
     <application
+        android:name="com.example.my_project.project.application.MyApplication"
         android:allowBackup="true"
         android:icon="@drawable/img_note"
         android:label="@string/app_name"
@@ -10,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.CurrentNote">
         <activity
-            android:name="com.example.myproject.project.MainActivity"
+            android:name="com.example.my_project.project.MainActivity"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -18,6 +19,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+
+
     </application>
+
 
 </manifest>
