+++ app/src/main/java/com/example/my_project/project/model/DataModel.kt
@@ -1,0 +1,11 @@
+package com.example.my_project.project.model
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+open class DataModel : ViewModel() {
+    val imageUri: MutableLiveData<String>  by lazy {
+        MutableLiveData<String>()
+    }
+
+}
