+++ app/src/main/java/com/example/my_project/project/util/OnBackPressedListener.kt
@@ -1,0 +1,5 @@
+package com.example.my_project.project.util
+
+interface OnBackPressedListener {
+    fun onBackPressed() :Boolean
+}
