+++ app/src/main/java/com/example/my_project/project/widget/WidgetNote.kt
@@ -1,0 +1,28 @@
+package com.example.my_project.project.widget
+
+import android.appwidget.AppWidgetManager
+import android.appwidget.AppWidgetProvider
+import android.content.Context
+
+class WidgetNote: AppWidgetProvider() {
+
+    override fun onUpdate(
+        context: Context?,
+        appWidgetManager: AppWidgetManager?,
+        appWidgetIds: IntArray?
+    ) {
+        super.onUpdate(context, appWidgetManager, appWidgetIds)
+    }
+
+    override fun onDeleted(context: Context?, appWidgetIds: IntArray?) {
+        super.onDeleted(context, appWidgetIds)
+    }
+
+    override fun onEnabled(context: Context?) {
+        super.onEnabled(context)
+    }
+
+    override fun onDisabled(context: Context?) {
+        super.onDisabled(context)
+    }
+}
