--- app/src/main/java/com/example/myproject/project/model/DataModel.kt
@@ -1,11 +1,0 @@
-package com.example.myproject.project.model
-
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-open class DataModel : ViewModel() {
-    val imageUri: MutableLiveData<String>  by lazy {
-        MutableLiveData<String>()
-    }
-
-}
