--- app/src/main/java/com/example/myproject/project/util/Constants.kt
@@ -1,14 +1,0 @@
-package com.example.myproject.project.util
-
-object Constants {
-
-    const val MAIN_REQUEST_CODE = 111
-    const val SHARED_PREF_KEY_NOTES_FRAGMENT = "grid variant"
-    const val SHARED_PREF_NAME_NOTES_FRAGMENT = "recyclerView state"
-    const val PARAMS_DETAIL_FRAGMENT = "DetailFragment key"
-    const val NOTES_FRAGMENT = "NotesFragment"
-    const val HIDDEN_FRAGMENT = "HiddenFragment"
-    const val PARAMS_PASSWORD_FRAGMENT = "PasswordFragment key"
-    const val DATE_FORMAT = "yyyy MM dd HH:mm:ss EE"
-
-}
