--- app/src/main/java/com/example/myproject/project/util/OnBackPressedListener.kt
@@ -1,5 +1,0 @@
-package com.example.myproject.project.util
-
-interface OnBackPressedListener {
-    fun onBackPressed() :Boolean
-}
