+++ app/src/main/res/layout/fragment_notes.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.myproject.project.list.NotesFragment"
+    tools:context="com.example.my_project.project.list.NotesFragment"
     tools:menu="@menu/notes_toolbar_menu">
 
     <LinearLayout
